Set up tsdoc markdown generation
It seems we manually need to run api-extractor before running
api-documenter; hence the bunch of configuration files.

The packages each get their own documentation in packages/*/api/docs (I’m
not sure we  need it per-package; if it could simplify configuration we
could just as well put it all into a single folder).

Ideally, the 'apache-annotator' meta-package would contain the
documentation of all packages it exports. The "bundledPackages" option
seems to be made for this, but when used I get a ‘not implemented’ error
because of the `import * as … from …` syntax we use. (I already had to
modify the original `export * as … from …` because of a similar error)
I suppose we could leave it for the time being, and be satisfied with the
documentation of the individual (non-meta) packages.